drone_sw/requirements.txt
adafruit-extended-bus
adafruit-circuitpython-lsm303-accel
adafruit-circuitpython-lsm303dlh-mag
micropython-ms5611
pynmea2
pyserial
orjson